+++ packages/apollo-angular-boost/src/index.ts
@@ -1,0 +1,12 @@
+import gql from 'graphql-tag';
+
+export * from 'apollo-client';
+export * from 'apollo-link';
+export * from 'apollo-cache-inmemory';
+export {Apollo, QueryRef} from 'apollo-angular';
+export {PresetConfig} from './types';
+export {APOLLO_BOOST_CONFIG} from './tokens';
+export {ApolloBoost} from './ApolloBoost';
+export {ApolloBoostModule} from './ApolloBoostModule';
+export {InMemoryCache} from 'apollo-cache-inmemory';
+export {gql};