+++ packages/apollo-angular-link-http-batch/src/index.ts
@@ -1,1 +1,3 @@
-export * from './api';
+export {HttpBatchLink, HttpBatchLinkHandler} from './HttpBatchLink';
+export {HttpBatchLinkModule, PROVIDERS} from './HttpBatchLinkModule';
+export {BatchOptions} from './types';