+++ packages/apollo-angular-link-http-common/src/index.ts
@@ -1,1 +1,2 @@
-export * from './api';
+export * from './utils';
+export * from './types';