+++ packages/apollo-angular/src/index.ts
@@ -1,1 +1,6 @@
-export * from './api';
+export {Apollo, ApolloBase} from './Apollo';
+export {QueryRef} from './QueryRef';
+export {Query} from './Query';
+export {SelectPipe} from './SelectPipe';
+export {ApolloModule} from './ApolloModule';
+export {APOLLO_OPTIONS} from './tokens';